updating a lot of file

+++ libraries/utilities/utilities.h
@@ -55,7 +55,7 @@ inline int Choose(int x, int y)
   return choose;
 }
 
-inline int Factorial(int x)
+inline double Factorial(int x)
 {
   return gsl_sf_fact(x);
 }
